Move styles to a separate files
