+++ src/vscode-form-control/vscode-form-control.styles.ts
@@ -1,0 +1,14 @@
+import {css, CSSResultGroup} from 'lit';
+import defaultStyles from '../includes/default.styles';
+
+const styles: CSSResultGroup = [
+  defaultStyles,
+  css`
+    :host {
+      display: block;
+      margin-top: 9px;
+    }
+  `,
+];
+
+export default styles;
