+++ src/vscode-form-description/vscode-form-description.ts
@@ -1,59 +1,14 @@
-import {css, CSSResultGroup, html, TemplateResult} from 'lit';
+import {html, TemplateResult} from 'lit';
 import {customElement} from 'lit/decorators.js';
-import defaultStyles from './includes/default.styles';
-import {VscElement} from './includes/VscElement';
+import {VscElement} from '../includes/VscElement';
+import styles from './vscode-form-description.styles';
 
+/**
+ * @deprecated
+ */
 @customElement('vscode-form-description')
 export class VscodeFormDescription extends VscElement {
-  static get styles(): CSSResultGroup {
-    return [
-      defaultStyles,
-      css`
-        :host {
-          color: var(--vscode-foreground);
-          cursor: default;
-          display: block;
-          font-family: var(--vscode-font-family);
-          font-size: var(--vscode-font-size);
-          line-height: 1.4;
-          margin: 3px 0;
-          opacity: 0.9;
-          overflow: hidden;
-          user-select: text;
-        }
-
-        ::slotted(a),
-        ::slotted(a:visited) {
-          color: var(--vscode-textLink-foreground);
-          text-decoration: none;
-        }
-
-        ::slotted(a:hover),
-        ::slotted(a:active) {
-          color: var(--vscode-textLink-activeForeground);
-          text-decoration: underline;
-        }
-
-        ::slotted(code) {
-          color: var(--vscode-textPreformat-foreground);
-        }
-
-        ::slotted(p) {
-          line-height: 1.4;
-          margin-bottom: 9px;
-          margin-top: 9px;
-        }
-
-        ::slotted(p:first-child) {
-          margin-top: 0;
-        }
-
-        ::slotted(p:last-child) {
-          margin-bottom: 0;
-        }
-      `,
-    ];
-  }
+  static styles = styles;
 
   render(): TemplateResult {
     return html` <slot></slot> `;
